requirements.txt
Pillow
scikit-image